Google Map Plugin
This is a WordPress plugin that generates a Google Map on a page with
the slug "map" wherever the shortcode [map] appears. The data from the
map comes from non-heirarchical (aka tag-like) custom taxonomies, also
generated by the plugin, for countries, states and cities. Add this
data to posts and the plugin generates a blank map with a "Show where
we've been" button that places markers on the map for each term. This
is a pre-1.0 release.